moved thread_management to proper package
