+++ src/core/hle/libraries/libkernel/thread_management.h
