+++ src/core/linker.cpp
@@ -4,7 +4,7 @@
 #include "common/string_util.h"
 #include "core/aerolib/aerolib.h"
 #include "core/aerolib/stubs.h"
-#include "core/hle/kernel/thread_management.h"
+#include "core/hle/libraries/libkernel/thread_management.h"
 #include "core/linker.h"
 #include "core/virtual_memory.h"
